Fix missing currency code
If you include the currency code in the Fingerprint you MUST include the
currency code in the POST request to the transaction gateway, or the
fingerprints will fail.

Notice that when using the Fingerprint tool
https://developer.authorize.net/tools/responsecode99/?page_id=7199
It will ask you for the currency code. The currency code is 3 characters
so make sure you are sending the correct currency code. The tool or
service does not issue an error.

That is if you use US instead of USD, your fingerprints will fail and
you will have no idea why. Same if you include the currency code in your
fingerprint and not int he PSOT request, the fingerprint will fail.