Remove currency code
There are apparently some issues with this. According to the
documentation, if you include a currency code in the hash, you need to
include the x_currency_code field in the form, however, no transactions
will go through with this.

We will just let it default to the merchant configured currency in the
settings of the authorize.net account.